backend/requirements.txt
# Python 3.11+ required
fastapi==0.115.0
uvicorn[standard]>=0.31.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic-settings>=2.4.0
python-multipart==0.0.12
pdfplumber==0.11.4
openpyxl==3.1.5
python-docx==1.1.2
# agent-framework - commented out due to dependency conflicts, will install manually if needed
# agent-framework>=1.0.0rc4
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
